+++ apps/laboratory/src/components/Wagmi/WagmiModalInfo.tsx
@@ -9,8 +9,8 @@ export function WagmiModalInfo() {
   const [clientId, setClientId] = React.useState<string | null>(null)
 
   async function getClientId() {
-    if (connector?.type === 'walletConnect' && typeof connector?.getProvider === 'function') {
-      const provider = await connector?.getProvider()
+    if (connector?.type === 'walletConnect') {
+      const provider = await connector?.getProvider?.()
       const ethereumProvider = provider as EthereumProvider
 
       return ethereumProvider?.signer?.client?.core?.crypto?.getClientId()