+++ packages/wallet/src/W3mFrameProvider.ts
@@ -1,8 +1,15 @@
 import { W3mFrame } from './W3mFrame.js'
 import type { W3mFrameTypes } from './W3mFrameTypes.js'
-import { W3mFrameConstants, W3mFrameRpcConstants } from './W3mFrameConstants.js'
+import { DEFAULT_LOG_LEVEL, W3mFrameConstants, W3mFrameRpcConstants } from './W3mFrameConstants.js'
 import { W3mFrameStorage } from './W3mFrameStorage.js'
 import { W3mFrameHelpers } from './W3mFrameHelpers.js'
+import {
+  generateChildLogger,
+  generatePlatformLogger,
+  getDefaultLoggerOptions,
+  type ChunkLoggerController,
+  type Logger
+} from '@walletconnect/logger'
 
 // -- Types -----------------------------------------------------------
 type Resolver<T> = { resolve: (value: T) => void; reject: (reason?: unknown) => void } | undefined
@@ -71,11 +78,40 @@ export class W3mFrameProvider {
 
   private setPreferredAccountResolver: SetPreferredAccountResolver = undefined
 
+  public logger: Logger
+
+  public chunkLoggerController: ChunkLoggerController | null
+
   public constructor(projectId: string) {
+    const loggerOptions = getDefaultLoggerOptions({
+      level: DEFAULT_LOG_LEVEL
+    })
+
+    const { logger, chunkLoggerController } = generatePlatformLogger({
+      opts: loggerOptions
+    })
+    this.logger = generateChildLogger(logger, this.constructor.name)
+    this.chunkLoggerController = chunkLoggerController
+
+    if (typeof window !== 'undefined' && this.chunkLoggerController?.downloadLogsBlobInBrowser) {
+      // @ts-expect-error any
+      if (!window.dowdownloadAppKitLogsBlob) {
+        // @ts-expect-error any
+        window.downloadAppKitLogsBlob = {}
+      }
+      // @ts-expect-error any
+      window.downloadAppKitLogsBlob['sdk'] = () => {
+        if (this.chunkLoggerController?.downloadLogsBlobInBrowser) {
+          this.chunkLoggerController.downloadLogsBlobInBrowser({
+            projectId
+          })
+        }
+      }
+    }
+
     this.w3mFrame = new W3mFrame(projectId, true)
     this.w3mFrame.events.onFrameEvent(event => {
-      // eslint-disable-next-line no-console
-      console.log('💻 received', event)
+      this.logger.info({ event }, 'Event received')
 
       switch (event.type) {
         case W3mFrameConstants.FRAME_CONNECT_EMAIL_SUCCESS: