+++ packages/polyfills/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @web3modal/polyfills
 
+## 5.0.3
+
+### Patch Changes
+
+- fix: ethers5 coinbase issues. Turbo build issues. Upate cb connector.
+
+## 5.0.2
+
+### Patch Changes
+
+- fix: siwe signOutOnNetwork change issue. fix: wallets filtered by rdns matched from explorer api. fix: solana network id issue
+
 ## 5.0.1
 
 ### Patch Changes